backend/requirements.txt

fastapi>=0.104.1,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6
pydantic>=2.5.0,<3.0.0
python-dotenv>=1.0.0
google-cloud-bigquery>=3.13.0
google-generativeai>=0.3.2
google-cloud-aiplatform>=1.38.1
PyPDF2>=3.0.1
python-docx>=1.1.0
jira>=3.5.2
httpx>=0.25.2
pydantic-settings>=1.2.1

# Vector database dependencies
langchain-postgres>=0.0.6
psycopg[binary]>=2.9.6
langchain>=0.0.352
langchain-google-genai>=0.0.6
psycopg2-binary>=2.9.6
cloud-sql-python-connector[asyncpg]
asyncpg
sqlalchemy
pg8000